(Refact) seperate chart utils from dailyChart.tsx
